requirements.txt
Pillow>=10.1.0
tk>=0.1.0
numpy>=1.26.2
